Change on castxml download script

+++ cmake/scripts/download_castxml_binaries.sh
@@ -20,24 +20,26 @@ if ! [ -d $1/castxml ] ; then
   axel_worked=0
   $2 -E make_directory $1 >/dev/null
   with_axel=$($2 -E echo $3 | grep -o "WITH_AXEL")
-  # Go to wget/curl if axel is not present
-  if [ ! -z "${with_axel}" ]; then
-    if command -v axel >/dev/null; then
-      if $2 -E chdir $1 axel --output=$1/$5 $4; then
-        axel_worked=1
-      else
-        $2 -E echo "Axel failed! The link probably redirects to https. Falling back to wget/curl..."
+  if [ ! -f $1/castxml-linux.tar.gz ]; then
+    # Go to wget/curl if axel is not present
+    if [ ! -z "${with_axel}" ]; then
+      if command -v axel >/dev/null; then
+        if $2 -E chdir $1 axel --output=$1/$5 $4; then
+          axel_worked=1
+        else
+          $2 -E echo "Axel failed! The link probably redirects to https. Falling back to wget/curl..."
+        fi
       fi
     fi
-  fi
-  if [ "${axel_worked}" = "0" ]; then
-    if command -v wget >/dev/null; then
-      wget --output-document=$1/$5 $4 
-    elif command -v curl >/dev/null; then
-      $2 -E chdir $1 curl -L -o $1/$5 $4
-    else
-      $2 -E cmake_echo_color --red --bold "ERROR: No axel, no wget, no curl?  What kind of OS are you running anyway?"
-      exit 1
+    if [ "${axel_worked}" = "0" ]; then
+      if command -v wget >/dev/null; then
+        wget --output-document=$1/$5 $4 
+      elif command -v curl >/dev/null; then
+        $2 -E chdir $1 curl -L -o $1/$5 $4
+      else
+        $2 -E cmake_echo_color --red --bold "ERROR: No axel, no wget, no curl?  What kind of OS are you running anyway?"
+        exit 1
+      fi
     fi
   fi
